Rebase from 'release/melodic/mrpt2'

+++ libs/opengl/shaders/points.f.glsl
@@ -1,11 +1,11 @@
 R"XXX(#version 300 es
 
 // FRAGMENT SHADER: Default shader for MRPT CRenderizable objects
-// Jose Luis Blanco Claraco (C) 2019-2020
+// Jose Luis Blanco Claraco (C) 2019-2022
 // Part of the MRPT project
 
-in highp vec4 frag_color;
-out highp vec4 color;
+in mediump vec4 frag_color;
+out mediump vec4 color;
 
 void main()
 {
